Fix module_inception
